test(textarea): recategorize and clean up browser tests
